Refactor logging structure: move LoggingFactory to new logger module and remove obsolete logger files
