+++ bipedal_robot_pkg/ros_nodes/logger/logging_factory.py
